refactor(gateway): extract events status collector module (#3212)
